+++ backend/app/services/catalog/material_hypothesis_service.py
