+++ backend/app/services/catalog/material_service.py
