+++ backend/app/services/floorplan/geometry_service.py
