+++ backend/app/services/floorplan/topology_service.py
