+++ backend/app/services/floorplan/wall_extraction_helpers/line_detection.py
