+++ backend/app/services/floorplan/wall_extraction_helpers/ocr.py
