+++ backend/app/services/rf/rf_run_service.py
@@ -94,7 +94,7 @@ async def create_rf_run(
 
     # 새 흐름: SageMaker async 호출
     if payload.access_points and payload.simulation:
-        from app.services.rf_job_service import submit_rf_simulation
+        from app.services.rf.rf_job_service import submit_rf_simulation
 
         rf_run, job = await submit_rf_simulation(
             db,
