+++ backend/app/services/scene/draft_object_service.py
