+++ backend/app/services/scene/object_service.py
@@ -14,7 +14,7 @@ from app.models.scene_version import SceneVersion
 from app.models.user import User
 from app.schemas.scene_object import ObjectCreate, ObjectResponse, ObjectUpdate
 from app.services._patch_log_helpers import record_patch, snapshot_object
-from app.services.scene_version_service import (
+from app.services.scene.scene_version_service import (
     _get_owned_scene_version,
     _object_to_response,
 )
