+++ backend/app/services/scene/opening_service.py
@@ -15,7 +15,7 @@ from app.models.user import User
 from app.models.wall import Wall
 from app.schemas.opening import OpeningCreate, OpeningResponse, OpeningUpdate
 from app.services._patch_log_helpers import record_patch, snapshot_opening
-from app.services.scene_version_service import (
+from app.services.scene.scene_version_service import (
     _get_owned_scene_version,
     _opening_to_response,
 )
