+++ backend/app/tests/test_logic.py
@@ -5,8 +5,8 @@ backend_dir = Path(__file__).resolve().parent.parent.parent
 sys.path.insert(0, str(backend_dir))
 
 from app.schemas.ai_output import MlOutputDTO
-from app.services.fusion_service import fusion_service
-from app.services.geometry_service import GeometryService
+from app.services.floorplan.fusion_service import fusion_service
+from app.services.floorplan.geometry_service import GeometryService
 
 def test_wi_twin_logic():
     print(" Wi-Twin 백엔드 통합 로직 테스트 시작")
